+++ src/lib.rs
@@ -1,7 +1,3 @@
-pub mod abe;
-#[cfg(test)]
-pub mod abe_tests;
-#[cfg(all(not(target_arch = "wasm32"), not(windows), feature = "libsodium"))]
 pub mod aes_hash_mmo;
 pub mod asymmetric;
 pub mod brc_c_prf_hi;
@@ -18,5 +14,6 @@ pub mod cs_prng;
 #[cfg(all(not(target_arch = "wasm32"), not(windows)))]
 pub mod primes;
 
+#[allow(deref_nullptr)]
 #[cfg(all(not(target_arch = "wasm32"), not(windows), feature = "libsodium"))]
 pub mod sodium_bindings;